new class Smart for update logic
